Refactor the matrix headers to separate concerns
