+++ glucat/glucat_imp.h
@@ -47,10 +47,6 @@
 
 #include "glucat/matrix_imp.h"
 
-#if defined(_GLUCAT_USE_ARMADILLO)
-#include "glucat/matrix_arma_imp.h"
-#endif
-
 #include "glucat/generation_imp.h"
 
 #include "glucat/matrix_multi_imp.h"
